+++ src/applications/vaos/appointment-list/components/ConfirmedAppointmentDetailsPage/index.jsx
@@ -5,7 +5,7 @@ import { useParams } from 'react-router-dom';
 import moment from 'moment';
 import ErrorMessage from '../../../components/ErrorMessage';
 import FullWidthLayout from '../../../components/FullWidthLayout';
-import VideoLayout from '../../../components/layout/VideoLayout';
+import VideoLayout from '../../../components/layouts/VideoLayout';
 import { selectFeatureBreadcrumbUrlUpdate } from '../../../redux/selectors';
 import {
   isAtlasVideoAppointment,
@@ -26,7 +26,7 @@ import {
 } from '../../redux/selectors';
 import PageLayout from '../PageLayout';
 import DetailsVA from './DetailsVA';
-import CCLayout from '../../../components/layout/CCLayout';
+import CCLayout from '../../../components/layouts/CCLayout';
 
 export default function ConfirmedAppointmentDetailsPage() {
   const dispatch = useDispatch();
