+++ src/applications/vaos/appointment-list/pages/CancelAppointmentPage/CancelPageLayoutRequest.jsx
@@ -8,7 +8,7 @@ import FacilityPhone from '../../../components/FacilityPhone';
 import {
   CCDetails,
   Details,
-} from '../../../components/layout/DetailPageLayout';
+} from '../../../components/layouts/DetailPageLayout';
 import NewTabAnchor from '../../../components/NewTabAnchor';
 import Section from '../../../components/Section';
 import { getRealFacilityId } from '../../../utils/appointment';
