+++ src/applications/vaos/appointment-list/pages/RequestedAppointmentDetailsPage/RequestedAppointmentDetailsPage.jsx
@@ -7,8 +7,8 @@ import ErrorMessage from '../../../components/ErrorMessage';
 import FacilityAddress from '../../../components/FacilityAddress';
 import FacilityPhone from '../../../components/FacilityPhone';
 import FullWidthLayout from '../../../components/FullWidthLayout';
-import CCRequestLayout from '../../../components/layout/CCRequestLayout';
-import VARequestLayout from '../../../components/layout/VARequestLayout';
+import CCRequestLayout from '../../../components/layouts/CCRequestLayout';
+import VARequestLayout from '../../../components/layouts/VARequestLayout';
 import VAFacilityLocation from '../../../components/VAFacilityLocation';
 import { selectFeatureBreadcrumbUrlUpdate } from '../../../redux/selectors';
 import { FETCH_STATUS } from '../../../utils/constants';
