+++ src/applications/vaos/components/layouts/VideoLayoutAtlas.unit.spec.js
