+++ src/applications/vaos/components/layouts/VideoLayoutVA.unit.spec.js
